Redisgned `modules/acmbot.py` decoupled code
A lot of functionality has been moved out of the AcmBotModule class.

* The loading of `events.yaml` is now done with the `load_yaml`
function.
* There are now filter functions which are responsible for sorting
and selecting which events are wanted. There filters are built opon
the select_events and sort_events functions
    * `select_events`
    * `sort_events`
    * `filter_take_date_range`
    * `filter_take_date`
* There are now command functions which implement the verious things
the acmbot can do.
    * `command_date`
    * `command_today`
    * `command_tomorrow`
    * `command_next`
    * `command_events`
    * `command_day`
* utility functions have been added
    * `weekday_difference`

The AcmBotModule is now responsible for responding to irc events,
loading configuration, calling global function above, constructing
messages, and responding to the irc event.